publisher/requirements.txt
pika
pydantic
python-dotenv
google-api-python-client
google-auth-oauthlib
